Refactor Carousel.tsx placement
