+++ Obligatory-Assignment-1/client/Obligatory-Assigment-1/src/components/Carousel/Carousel.tsx
