+++ Obligatory-Assignment-1/client/Obligatory-Assigment-1/src/pages/Home.tsx
@@ -1,4 +1,4 @@
-import Carousel from "../components/Carousel.tsx";
+import Carousel from "../components/Carousel/Carousel.tsx";
 import CarouselPicture1 from "../components/Carousel/Carousel_Assets/CarouselPicture1.png";
 import CarouselPicture2 from "../components/Carousel/Carousel_Assets/CarouselPicture2.png";
 import CarouselPicture3 from "../components/Carousel/Carousel_Assets/CarouselPicture3.png";
